Added command line for Windows
There was only the sudo version, so I added the sudo-less command for Windows users, clarified the OSes, etc. Note separately that the Wiki troubleshooting page (https://github.com/npm/npm/wiki/Troubleshooting) is mistaken about the paths in Windows. I can't find a way to edit that page, so can someone on the team fix it? The node installer puts the "...Program Files/node.exe" path in System Paths and puts the npm "...AppData/Roaming/npm..." path in User paths. One is not "before" the other – they're in separate environment variables. This is true on Windows 8.1 and 10.